began writing mouse smoothening algorithim in rust

+++ go-rust-backend/internal/video/video-editing-engine/video-effects-processor/include/video_editing_engine.h
@@ -6,9 +6,9 @@
 #include <stdio.h>
 
 typedef struct {
-  double x;
-  double y;
-  int64_t timestamp_ms;
+  int x;
+  int y;
+  double timestamp_ms;
 } CPoint;
 
 typedef struct {
